Introducing VSS Enterprise Dashboard
Added 3 queries, 3 visualizations to complete graphing requirement
for VSS-Enterprise Dashboard.

Note: Temporarily I have kept vss-top-sec-events as vertical bar chart
but as soon as the pie-chart component is ready, we need to make a change
there. Have verified that the response is coming properly for all 3 queries and
for first 2 (acl-deny over time and top-x-domains) the graph is showing correctly.