Adding Enterprise level metrics
Using SimpleTextGraph for now to bring in 3 metrics at enterprise level.
Idea is to show the total count for a metric (tca-events, acl-deny and flows/packets)
for last-24 hours as well the positive or negative difference compare to prev 24 hours.
Query is been validated. I need some help with a modified SimpleTextGraph to show this
information better.