Top 5 upload/download query correction
1. Upload
Changing the aggregation based on bytes iso packets.

2. Download
a.Changing the aggregation based on bytes iso packets.
b.Current PPS/AAR feature behavior is that when the packet
comes from network to access it is being reported as an ingress from the
destination-NSG. So query need to be made accordingly.